requirements.txt
requests
beautifulsoup4
python-dotenv
google-genai
